final commit

+++ src/main/java/com/retail/retaildiscountapp/model/IUser.java
@@ -4,5 +4,5 @@ public interface IUser {
 	
 	public boolean isEmployee();
 	public boolean isAffiliate();
-	public long getMemberSinceYears();	
+	public int getMemberSinceYears();	
 }
